backend/requirements/requirements-production.txt
Django>=3.1
psycopg2-binary
django-filter
djangorestframework
requests
python3-memcached
gunicorn
django-admin-tools
djangorestframework-xml
uritemplate
django-cors-headers
redis
channels
channels_redis
twilio
django-leaflet
